Version 0.2.0: Add caching + update to new NHL api
Most important change was to read from the new NHL api.
Te NHL changed their API mid season, so that caused the countdown
server to stop working completely. I decided to use this chance to
also rewrite the application for better performance, generally,
and more specifically, it now caches the NHL's responses, rather
than doing a GET request to the NHL server each request this
server got.

The caching implementation creates a static value for each team,
that value is only created on the first request. So we basically
map each team to a static location in memory that each request
can read safely read from in parallel. RwLock is used for each
location, so it can be updated when needed. This is more
performant than using a RwLock on a hashmap, because each team
has its own lock. Thus when the schedule for one team changes, the
other requests can just read their team immediatly, without having
to wait on the data for another team to update.

The logic after the rewrite for the app is we get a `Game` type
for the current or next game. That type also handles what is
displayed on the lametric. Because this type is also just a
simplified form of the data we get from the NHL api, there is a
very simple and clean way to add more data to the lametric output,
say if we wanted to print the scores while in game :)

Also, general dependency updates, and removal of deprecated code.